Developer/requirements.txt
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
moviepy>=1.0.3
edge-tts>=6.1.9
tqdm>=4.65.0
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0 